splitting db from the root repo
