+++ src/db/index.html
@@ -7,6 +7,7 @@
 </head>
 
 <body>
+    <script src="./index.js"></script>
     <script src="./process.js"></script>
     <div>
         <div>
