+++ src/db/package.json
@@ -1,0 +1,40 @@
+{
+    "name": "unbundled",
+    "version": "0.0.1",
+    "description": "This is an attempt at unbundling the web.",
+    "main": "index.js",
+    "scripts": {
+        "start": "parcel --no-cache ./index.html",
+        "test": "echo \"Error: no test specified\" ",
+        "build": "parcel build ./index.html"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/humphd/next.git"
+    },
+    "author": "Dmytro Sych",
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/humphd/next/issues"
+    },
+    "homepage": "https://github.com/humphd/next#readme",
+    "devDependencies": {
+        "babel-plugin-transform-runtime": "^6.23.0",
+        "parcel-bundler": "^1.8.1",
+        "parcel-plugin-sw-cache": "dsych/parcel-plugin-sw-cache#relative-dest"
+    },
+    "dependencies": {
+        "dexie": "^2.0.3"
+    },
+    "cache": {
+        "strategy": "inject",
+        "globDirectory": "dist/",
+        "globPatterns": [
+            "**/*.{js,html}"
+        ],
+        "swSrc": "dist/db-worker.js",
+        "swDest": "dist/db-worker.js",
+        "useRelativeDest": true,
+        "inDev": true
+    }
+}
