--- src/index.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <script src="./index.js"></script>
-</head>
-
-<body>
-    <h1>Welcome to Unbundled</h1>
-    <p>
-        Make something
-        <strong>amazing</strong> with the web!
-    </p>
-
-    <a href="./db/index.html">DB interface</a>
-</body>
-
-</html>
